requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlmodel>=0.0.14
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
httpx>=0.26.0
